Shell file creation and partial completion
Created the shell script which allows the challenge to be run on Linux terminals with partial completion such as key randomisation (which will be used to encrypt the files) and file creation.

The remaining aspects is the win message when the user reaches the final file and echoing the partial key of the next file into the previous one.