Shell script completion
Completed the remaining aspects of the shell script which is the win message and echoing the partial keys. As the purpose of the project is to challenge the user to decrypt each file in a numerical order, each file will hold the partial key for the next file, but get its current one held by the previous (which for the first file gets echoed to the terminal to help the user).

Also changed the structure of the generateKey by adding the potential of the key being a SHA1 hash and using the partial key variable as an argument (which is setup in the challangeSetup method), as otherwise the partial key would incorrectly return a number and only be accessible after the function is called.