Dash removal and shell decryptor
When a key is created with a specific hash sum (such as MD5) the value includes a - (dash) at the end, making it difficult for the user to decrypt the file as ccdecrypt wouldn't recognise it as part of the full key.

To avoid this, the addition of the cut command ensures that no matter which hash/string is chosen to generate the key with, there will be no dashes included.

Also added a decrypter script (for the shell version) to allow the user to pass in the file name and its key instead of repeatedly calling the ccdecrypt command and passing in arguments.